requirements.txt
Pillow==11.3.0
colorama==0.4.6
pyinstaller==6.14.2